docs_src/requirements.txt
codetiming
memory_profiler
numpy<2.0
pandas
pytest
torch==2.0.1
torchvision==0.15.2
transformers
tqdm
shibuya
sphinx-copybutton
sphinxemoji
sphinx-sitemap
sphinx_contributors
sphinx_docsearch